rfctr: remove smoother from cube module
